ONB-197 refactor: logger() 이동
