--- adapter-in/web/src/main/kotlin/com/yapp/bol/utils/Logger.kt
@@ -1,5 +1,0 @@
-package com.yapp.bol.utils
-
-import org.slf4j.LoggerFactory
-
-inline fun <reified T> T.logger() = LoggerFactory.getLogger(T::class.java)
