+++ support/logging/src/main/kotlin/com/yapp/bol/logging/utils/Logger.kt
@@ -1,0 +1,5 @@
+package com.yapp.bol.logging.utils
+
+import org.slf4j.LoggerFactory
+
+inline fun <reified T> T.logger() = LoggerFactory.getLogger(T::class.java)
